FinalProject
Created a new console application and created two extra files: BlingBlong.cs and Sequence.cs. I then created a namespace, class and method in BlingBlong that ran through the numbers 1 to 100 and any number that was divisible by 3 was replaced with the word Bling and any number that was divisible by 5 was replaced with Blong, and any number that was divisible by both was replaced with the word BlingBlong. In Sequence I wrote a simple code to print the Fibonacci sequence to the console.